Admin car uploads
The admin can now upload cars. PHP files done for it. The only thing that needs to be adjusted is the structure of the Javascript code due to security reasons.

The Javascript code needs to fetched via PHP. In order to access the Javascript code, the admin has to login and I need to attach the stored username with the fetch!