Loading animation, phone adjustments
A lot of adjustments have been done. Two of the most important ones have been the loading menu animation and the phone hamburger menu. They look clean af now.

What you need to do now is to make sure that the sizes of the "Latest Cars" are proper and check if they work.